Move the example to agbench GAIA templates
