+++ python/packages/agbench/benchmarks/GAIA/Templates/ParallelAgents/expected_answer.txt
@@ -1,0 +1,1 @@
+__EXPECTED_ANSWER__
