+++ python/packages/agbench/benchmarks/GAIA/Templates/ParallelAgents/prompt.txt
@@ -1,0 +1,1 @@
+__PROMPT__
