--- python/samples/parallel_agents/config.yaml
@@ -1,38 +1,0 @@
-# config.yaml
-#
-# The contents of this file will be copied into the 'config.yaml' file of
-# every expanded Task, just prior to running the scenario. This provides a
-# good place to store model or other configurations important for the scenario.
-
-###############################
-# Open AI model configuration #
-###############################
-model_config: &client
-  provider: autogen_ext.models.openai.OpenAIChatCompletionClient
-  config:
-    model: gpt-4o
-
-
-##############################
-# Ollama model configuration #
-##############################
-#model_config: &client
-#    provider: autogen_ext.models.openai.OpenAIChatCompletionClient
-#    config:
-#      model: deepseek-r1:7b
-#      base_url: http://localhost:11434/v1/
-#      api_key: ollama
-#      model_info:
-#        function_calling: false
-#        json_output: false
-#        vision: false
-#        family: r1
-#
-
-#######################
-# Used by MagenticOne #
-#######################
-orchestrator_client: *client
-coder_client: *client
-web_surfer_client: *client
-file_surfer_client: *client
