--- python/samples/parallel_agents/requirements.txt
@@ -1,84 +1,0 @@
-aiofiles==24.1.0
-annotated-types==0.7.0
-anyio==4.9.0
-audioop-lts==0.2.1
-autogen-agentchat==0.5.3
-autogen-core==0.5.3
-autogen-ext==0.5.3
-azure-ai-documentintelligence==1.0.2
-azure-core==1.33.0
-azure-identity==1.21.0
-beautifulsoup4==4.13.4
-certifi==2025.1.31
-cffi==1.17.1
-charset-normalizer==3.4.1
-click==8.1.8
-cobble==0.1.4
-coloredlogs==15.0.1
-cryptography==44.0.2
-defusedxml==0.7.1
-Deprecated==1.2.18
-distro==1.9.0
-et_xmlfile==2.0.0
-flatbuffers==25.2.10
-greenlet==3.2.0
-h11==0.14.0
-httpcore==1.0.8
-httpx==0.28.1
-humanfriendly==10.0
-idna==3.10
-importlib_metadata==8.6.1
-isodate==0.7.2
-jiter==0.9.0
-jsonref==1.1.0
-lxml==5.3.2
-magika==0.6.1
-mammoth==1.9.0
-markdownify==1.1.0
-markitdown==0.1.1
-mpmath==1.3.0
-msal==1.32.0
-msal-extensions==1.3.1
-numpy==2.2.4
-olefile==0.47
-onnxruntime==1.21.0
-openai==1.75.0
-openpyxl==3.1.5
-opentelemetry-api==1.32.1
-packaging==24.2
-pandas==2.2.3
-pdfminer.six==20250416
-pillow==11.2.1
-playwright==1.51.0
-protobuf==5.29.4
-pycparser==2.22
-pydantic==2.11.3
-pydantic_core==2.33.1
-pydub==0.25.1
-pyee==12.1.1
-PyJWT==2.10.1
-python-dateutil==2.9.0.post0
-python-dotenv==1.1.0
-python-pptx==1.0.2
-pytz==2025.2
-PyYAML==6.0.2
-regex==2024.11.6
-requests==2.32.3
-six==1.17.0
-sniffio==1.3.1
-soupsieve==2.6
-SpeechRecognition==3.14.2
-standard-aifc==3.13.0
-standard-chunk==3.13.0
-sympy==1.13.3
-tiktoken==0.9.0
-tqdm==4.67.1
-typing-inspection==0.4.0
-typing_extensions==4.13.2
-tzdata==2025.2
-urllib3==2.4.0
-wrapt==1.17.2
-xlrd==2.0.1
-XlsxWriter==3.2.3
-youtube-transcript-api==1.0.3
-zipp==3.21.0
