--- python/samples/parallel_agents/utils.py
@@ -1,45 +1,0 @@
-import json
-import logging
-from datetime import datetime
-from autogen_agentchat.utils import content_to_str
-from autogen_core.logging import LLMCallEvent
-from autogen_agentchat import EVENT_LOGGER_NAME as AGENTCHAT_EVENT_LOGGER_NAME
-from autogen_core import EVENT_LOGGER_NAME as CORE_EVENT_LOGGER_NAME
-
-class LogHandler(logging.FileHandler):
-    def __init__(self, filename: str = "log.jsonl", print_message: bool = True) -> None:
-        super().__init__(filename, mode="w")
-        self.print_message = print_message
-
-    def emit(self, record: logging.LogRecord) -> None:
-        try:
-            ts = datetime.fromtimestamp(record.created).isoformat()
-            # print(f"{ts} {record.levelname} {record.name}: {record.msg}", flush=True)
-            if AGENTCHAT_EVENT_LOGGER_NAME in record.name:
-                original_msg = record.msg
-                record.msg = json.dumps(
-                    {
-                        "timestamp": ts,
-                        "source": record.msg.source,
-                        "message": content_to_str(record.msg.content),
-                        "type": record.msg.type,
-                    }
-                )
-                super().emit(record)
-                record.msg = original_msg
-            elif CORE_EVENT_LOGGER_NAME in record.name:
-                if isinstance(record.msg, LLMCallEvent):
-                    original_msg = record.msg
-                    record.msg = json.dumps(
-                        {
-                            "timestamp": ts,
-                            "prompt_tokens": record.msg.kwargs["prompt_tokens"],
-                            "completion_tokens": record.msg.kwargs["completion_tokens"],
-                            "type": "LLMCallEvent",
-                        }
-                    )
-                    super().emit(record)
-                    record.msg = original_msg
-        except Exception:
-            print("error in logHandler.emit", flush=True)
-            self.handleError(record)
